AJ-1286: Introduce `BiStream` idioms using `com.google.mug` (#329)
While working on `DataTypeInferer` I noticed these stanzas of code which cried out for `BiStream` which was designed exactly for this sort of streaming logic.  See https://github.com/google/mug#bistream-streams-pairs-of-objects for many examples.

This commit is pure refactoring and is intended to demonstrate how `BiStream` can be used to simplify complex iterations and by using higher-level predicates.