AJ-1234: Relax regex matching conflict message.
The actual message differed slightly from expectation, so I adjusted
the regex to allow both forms of pluralization; what matters most is
that the conflict was due to something relating to policies (and not
one of the other myriad ways a 409 can be returned).

Also update the example message to match the (current) actual message
we expect.