Expose TwdsProperties child props as `@Bean`s.
This allows classes to autowire only the properties they need, rather
than getting _all_ the properties and having to traverse the hierarchy.

It also allows more rapid failure due to misconfigured beans, so we can
be less defensive about null checking and can assume for the most part
the beans will be present.