AJ-1650: Clean up properties. (#583)
Clean up `TwdsProperties`` and its children props classes to improve
fail-fast semantics and configuration for the WORKSPACE_ID environment
variable:

* If in single-tenant mode, WORKSPACE_ID is required, and WDS should
crash instead of allowing any code to run without it. This is enforced
by providing access to this environment variable through the
`InstanceProperties.workspaceId()` accessor, which is designed to never
return null. In fact, InstanceProperties can't be resolved as a Bean at
all unless it has a valid WorkspaceId, so this can be used to enforced
in depth when injected via InstanceProperties.
* A convenience `@Bean` for the `WorkspaceId`` annotated as
`@SingleTenant` is provided for direct injection, but still benefits
from the same constraints. Where possible, this is the easiest way to
get access to `WorkspaceId.`
* A couple of example changes are provided in the code & tests where
the prior approach of injecting `instanceProperties` needed to be
updated, and demonstrate how we might stamp this out more broadly.

Note: Direct access to the workspace-id using
`@Value('twds.instance.workspace-id')` throughout the codebase still
circumvents the validation and enforcement that it's non-null,
non-blank, and a valid UUID. Followup is required to eliminate these
references, which will probably cause unexpected misbehavior in
control-plane.