Clarify `HttpSamDao.hasWriteWorkspacePermission`
The single String parameter was named as though it were an auth token,
but all places (except one) where it was used (in tests and prod) it was
being treated as a workspaceId.  This clarifies the parameter name, and
wires it through correctly to be used for the auth check.