Extract `SamAuthorizationDao` interface.
This is a mostly mechanical IDE-driven refactor to extract the
permission-checking parts of `SamDao`'s API to a separate interface.

It's separated out into its own commit to isolate the changes that were
largely IDE driven (plus the manual corrections required to keep the
code compiling and passing tests).

`SamConfig` temporarily duplicates the creation code to provide
`@Bean` implementations for both interfaces.  This will be cleaned up
in a subsequent commit.