AJ-1633: Make `SamDao` API require and use stronger types. (#591)
Prefactor for [AJ-1633](https://broadworkbench.atlassian.net/browse/AJ-1633)

Strengthen `SamDao` API typing:
* Replace all parameters of `SamDao` that take a `String token` to
instead require `BearerToken`.
* For non-authenticated usages it's still possible to pass in
`BearerToken.empty()`.
* Replace all parameters of `SamDao` that take a `String workspaceId` to
instead require `WorkspaceId`.

Also, clarify `HttpSamDao.hasWriteWorkspacePermission`: the single
String parameter was named as though it were an auth token,
but all places (except one) where it was used (in tests and prod) it was
being treated as a workspaceId.  This clarifies the parameter name, and
wires it through correctly to be used for the auth check.