Split `SamAuthorizationDao` out from `SamDao`
The methods that remain on `SamDao` don't have an auth concern, which
requires a `WorkspaceId`.  Splitting these interfaces out will allow
two different implementations to replace `HttpSamDao`:
* `HttpSamDao` will be wired up with a `BearerToken` that can be used
instead of requiring it to be provided in the API methods.
* `HttpSamAuthorizationDao` will be wired up with both a `BearerToken`
and a `WorkspaceId`, allowing these parameters to be removed from the
API methods.  This will also allow us to eliminate the many overloads
of each method.
* `SamConfig` temporarily duplicates the creation code to provide
`@Bean` implementations for both interfaces.  This will be cleaned up
in a subsequent commit.