Extract `HttpSamAuthorizationDao`
* Extract `HttpSamAuthorizationDao` from `HttpSamDao` to fully divide
the class hierarchy implementing `SamAuthorizationDao` and `SamDao`.
* Rename `MisconfiguredSamDao` to `MisconfiguredSamAuthorizationDao`
and strip out the `SamDao` parts of the interface to reflect that it
only needs to exist for the `SamAuthorizationDao` concerns.  While doing
so, consolidate the implementation to placate Sonar warning: java:S4144
* Introduce `SamAuthorizationDaoFactory` to handle creation of the new
`HttpSamAuthorizationDao` classes.
Currently, it is just used to instantiate the `@Bean` class provided
by `SamConfig`, however, it is introduced to provide the following
future benefits:

1. Allow deferred creation so a request can create Dao lazily with
the appropriate context from the request.
2. For `SamAuthorizationDao`: inject `WorkspaceId` to eliminate it from
the API.
3. Eventually, for both `SamAuthorizationDao` and `SamDao`:
inject `BearerToken` to eliminate it from the API.