Extract `HttpSamAuthorizationDao`
* Extract `HttpSamAuthorizationDao` from `HttpSamDao` to fully divide
the class hierarchy implementing `SamAuthorizationDao` and `SamDao`.
* Rename `MisconfiguredSamDao` to `MisconfiguredSamAuthorizationDao`
and strip out the `SamDao` parts of the interface to reflect that it
only needs to exist for the `SamAuthorizationDao` concerns.  While doing
so, consolidate the implementation to placate Sonar warning: java:S4144
* Introduce `SamDaoFactory` to handle creation of the new Dao classes.
Currently, it is just used to instantiate the `@Bean` classes provided
by `SamConfig`, however, it is introduced to provide the following
future benefits:

1. Allow deferred creation so a request can create Daos lazily with
the appropriate context from the request.
2. For both Daos: inject `BearerToken` to eliminate it from the API.
3. For `SamAuthorizationDao`: inject `WorkspaceId` to eliminate it from
the API.