Fix `generatedSourceDirs` locations.
This was incorrectly labeling the `main` directory as a source root in
IDEA, which then included the 'java' directory as part of the source
code path.  I was noticing issues like partial package highlighting and
compilation errors in the IDE that went away when I changed this to more
accurately match the `src/main/java` convention.