AJ-1633 Disable workspaceId check for `LogStatementTest` (#624)
Disable workspaceId check for `LogStatementTest`

This test is failing because it is creating/encountering collections
that have a workspaceId that mismatches the `WORKSPACE_ID`.  Disabling
this check allows these tests to keep passing (and verifying what they
are actually interested in), without having to redesign them for the
concern of the safety check.