AJ-1669: Reuse `RecordSink` for full import. (#628)
This makes an entire import job use the same `RecordSink` instance for
all batches.  The goal here is to make the `RawlsRecordSink` capable of
handling the entire import job with a single GCS file & pubsub, rather
than triggering a new file & pubsub message for each batch.

In a subsequent change, the `RecordSink` will be given an additional
method to finalize the import, which is what will be called at the very
end of a batch to do all the post-processing.