Write and publish only once.
Goal: Rather than writing a Blob for every batch and every record type,
record all writes for a whole batch into one Blob and send a single
Pubsub message.

To this end:
* Add an E2E test with a file that exceeds the batch size to ensure the
multi-batch logic is triggered.
* Modify `GcsStorage` to provide an `OutputStream` instead of expecting
a `String` for a direct write.

Some cleanup along the way:
* `RecordSink` and subclasses: Rather than bubbling up checked
`IOException`s which callers are forced to handle, surface them as
`DataImportException`s which are treated as a server error.
* Switch `RecordSink` to implement `AutoCloseable` which is a more
general-purpose interface that doesn't imply a checked `IOException`.
* In Job classes, we can rewrap these exceptions as specific TDR/PFB
exceptions to provide more specificity.
* `RawlsRecordSink`: Clarify the instance variables used for pubsub with
an eye toward eventually pulling that concern out.