Rename `{BatchWrite->TypeMismatch}Exception`
This frees up the very generic `BatchWriteException` name so that it
doesn't create ambiguity when I introduce a new exception that might
be thrown during a batch write.

Also, eliminate all `throws` of this exception from the `RecordSink`
interface and its subclasses, since none of them ever actually throw it.