Prefer factory method for complex construction.
While the current construction of `GcsStorage` and `RawlsRecordSink`
are simple, the construction will both get more complicated in the next
commit.

In prep for this, instead of having potentially complicated logic that
potentially throws exceptions in the constructors, do this work in
factory methods (`#create`). This lets the constructors become really
simple and just set fields, which should improve unit testability and
set thnigs up for more complex construction.  It also isolates the
calling code from internal restructuring of the constructors which
is imminent.