AJ-1669: Write and publish only once. (#646)
Goal: Rather than writing a Blob for every batch and every record type,
record all writes for a whole batch into one Blob and send a single
Pubsub message.

* Rename `{BatchWrite->TypeMismatch}Exception`: this frees up the
very generic `BatchWriteException` name to disambiguate it from 
`DataImportException`.  Also, eliminate all `throws` of this exception
from the `RecordSink` interface and its subclasses, since none of
them ever actually throw it.
* `GcsStorage` and `RawlsRecordSink`: prefer factory method for
complex construction: instead of having potentially complicated logic
that potentially throws exceptions in the constructors, do this work in
factory methods (`#create`). This lets the constructors become really
simple and just set fields, which should improve unit testability and
set thnigs up for more complex construction.  It also isolates the
calling code from internal restructuring of the constructors which
happens in this PR.
* Make `RecordSink` implement `AutoCloseable`: This makes 
cleanup / finalizing automatic, whereas before it was easy to forget to 
call `finalizeBatchWrite` and leave resources open.  This wasn't
previously a problem because there were no resources being
held open by `RecordSink`, but after introducing statefulness this
problem became apparent in tests that began failing due to not calling
the cleanup method.
* Add an E2E test with a file that exceeds the batch size to ensure the
multi-batch logic is triggered.
* `RecordSink` and subclasses: Rather than bubbling up checked
`IOException`s which callers are forced to handle, surface them as
`DataImportException`s which are treated as a server error.
* In Job classes, rewrap these exceptions as specific TDR/PFB exceptions
to provide more specificity.
* `RawlsRecordSink`: Clarify the instance variables used for pubsub with
an eye toward eventually pulling that concern out.
* Modify `GcsStorage` to provide an `OutputStream` instead of expecting
a `String` for a direct write; this is the key functional change to allow
a single file to be written instead of multiple.