Change `QuartzJob` to label jobType with JobKey
This changes the metrics generated by `QuartzJob` to be labeled with the
JobKey.group, which is set to be the `ImportRequestServerModel.TypeEnum`
which will be either `PFB` or `TDRMANIFEST`.  These values are
potentially more stable than the class name (although that probably
won't change much), but more importantly, they are stored in the Job
`input` field, which is accessible later on during job status checks.

I plan to use the `input` field to label metrics I gather later on in
`processJobStatusUpdate` and I want the labels there to match what is
used up front in `QuartzJob`.