AJ-1726 Add measurement to job updates (#699)
Summary of changes:

* Measure `wds.job.elapsed`: the time elapsed since the job was created.
* Do metric labeling from  `QuartzJob` subclasses: this changes the
metrics generated by `QuartzJob` to be explicitly labeled by the
subclass implementations.
* Switch `jobType` label to actually be the `JobType`, which is based
on the `JobType` enum rather than the subclass name.  For now, this
value is only `DATA_IMPORT`.
* Add a new label called `importType` which is populated based on the
`TypeEnum` import type, which will be either `PFB` or `TDRMANIFEST`.
These values are potentially more stable than a class name (although
that probably won't change much), but more importantly, the `importType`
is stored in the Job `input` field, which is accessible later on during
job processing and can be used to label downstream metrics consistently.
* Rename `JobService{BaseTest} -> {TestBase}`: this rename clarifies
that the `TestBase` is the base class for all tests, not a test class.