AJ-1700: Fix bug while copying RAWLSJSON file.
Before this change, the `moveBlob` method would attempt to get the blob
based on the original target to return a reference to it.  However, the
original `BlobId` used as a reference also contained the original
`generation`.  Instead, we use the resultant `Blob` from the copy, which
has the correct new `generation`, and should actually exist.